Huge Update
After a long time, I took the time to improve this site. I put a lot of css, which was repeated in all css files, in global.css, which should improve the web loading speed. Using the tinypng tool (https://tinypng.com), I reduced the size of the images, which is also a great benefit for the speed of the site.